requirements.txt
flask
pillow
requests
arabic-reshaper
python-bidi